Fix issue with nameerrors being circumvented
